mv `SwapSegment` into operation
